+++ lib/editor/editor.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { ANGULAR_MONACO_EDITOR_CONFIG, AngularMonacoEditorConfig } from './model/config';
 import { AngularMonacoEditorComponent } from './angular-monaco-editor/angular-monaco-editor.component';
-import { AngularMonacoEditorService } from './service/angular-monaco-editor.service';
 
 @NgModule({
   imports: [
@@ -16,7 +15,6 @@ import { AngularMonacoEditorService } from './service/angular-monaco-editor.serv
     AngularMonacoEditorComponent
   ],
   providers: [
-    AngularMonacoEditorService
   ]
 })