+++ lib/editor/index.ts
@@ -5,7 +5,4 @@ export * from './model/types';
 
 export * from './constant/events';
 
-export * from './service/code-editor.event.service';
-export * from './service/angular-monaco-editor.service';
-
 export *  from './angular-monaco-editor/angular-monaco-editor.component';