+++ public_api.ts
@@ -5,6 +5,4 @@ export * from './lib/editor/model/types';
 
 export * from './lib/editor/constant/events';
 
-export * from './lib/editor/service/angular-monaco-editor.service';
-
 export * from './lib/editor/angular-monaco-editor/angular-monaco-editor.component';