+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 - [check editor's markers]
 - [demos to using angular-monaco-editor component]
 - [Init monaco editor using angular]
+- [Output onInit onBlurEditorText events]
 
+# 1.8.0
+  Record verified result when using schema; Add the onBlurEditorText output event
+  
 # 1.7.0
   Extract model as a service from the angular-monaco-editor component.